mcp_project/pyproject.toml
[project]
name = "mcp-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.60.0",
    "arxiv>=2.2.0",
    "groq>=0.30.0",
    "mcp>=1.12.3",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.1.1",
]
